+++ CHANGELOG.md
@@ -7,7 +7,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - added MAX/MIN/AVG to parameters line charts
 - added associated equipments trend data to combined equipment input energy category report
+- added new protocol to datasource in myems-api
+- added meter saving report
+- added meter comparison report
+- 
 ### Changed
+- updated demo database
+- updated docker build instruction in README
 - 
 ### Fixed
 -